simplify initial boot
